+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `requiresConfirmation` support for email form elements
+
 ## [8.3.0] - 2025-05-22
 
 ### Added